2回めの修正

+++ memo_app.rb
@@ -33,7 +33,7 @@ get '/memos/new' do
 end
 
 post '/memos' do
-  hash = { id: SecureRandom.uuid, title: params[:title], content: params[:content], time: Time.now.to_time.iso8601 }
+  hash = { id: SecureRandom.uuid, title: params[:title], content: params[:content], time: Time.now.iso8601 }
   File.open("data/#{hash[:id]}.json", 'w') { |file| JSON.dump(hash, file) }
   redirect '/memos'
 end
@@ -49,14 +49,18 @@ get '/memos/:id/edit' do
 end
 
 patch '/memos/:id' do
-  File.open(path_json_file, 'w') do |file|
-    hash = { id: params[:id], title: params[:title], content: params[:content], time: Time.now.to_time.iso8601 }
-    JSON.dump(hash, file)
+  if File.exist?(path_json_file)
+    File.open(path_json_file, 'w') do |file|
+      hash = { id: params[:id], title: params[:title], content: params[:content], time: Time.now.iso8601 }
+      JSON.dump(hash, file)
+    end
+  else
+    redirect to('not_found')
   end
   redirect '/memos'
 end
 
 delete '/memos/:id' do
-  File.delete(path_json_file)
+  File.exist?(path_json_file) ? File.delete(path_json_file) : (redirect to('not_found'))
   redirect '/memos'
 end
